Modified initialization function to split into separate parts
I originally only had the set long press function and was using during initialization. Under the assumption that it would only be needed once, I would reset the button after finishing.
However, I found that sometimes it is very useful to alter the long press of a button on the fly. For example having a 10 second long press during start up to enter a specific and then the normal 1-2 second long press off afterwards.
The set long press function would cause false button presses to register because I would reset the state of the button.
Now the user can choose to initialize the button again and reset the state, or just change the long press time. However, it would be wise to only change the long press time in a known state.