Fixed bug with GPIO implementation
Also I got rid of the direction type and just rolled it into type. That's because my previous thought was just restricted to Microchip, where direction is treated as a separate register than type (analog or digital). But when you think about it, I really don't need analog out and digital can be split into digital in and digital out. No I do not want to incorporate a DAC for analog out. That's causes more dependancies. GPIO libraries should only be concerned with the most basic GPIO functionality. Not alternate functions. That is what the interface is for. To allow inheritance via the implementation's header file.