Added UART_TransmitFinishedCheck to interface and UART1.
I realized that if the user is using interrupts and they call transmit from their callback (which they most likely will want to do) then it will create multiple recursive function calls.
The solution is to lock the transmit finished interrupt and set a flag. Then outside somewhere, the TransmitFinishedCheck function will look for this flag.
This should let the first TransmitFinishedEvent function finish and keep from pushing a bunch of these onto the stack
This issue will only every arise if the user is using transmit interrupts.