Rolled GPIO Pin List into GPIO Manager. Made GPIO Manager generic
The GPIO driver should only be concerned with handling pins. It shouldn't dictate how the user handles setting up all of their pins.
So once I moved this initialization and renamed it "Manager_STM32G0" I realized that it was basically generic at this point, and renamed it again.
Then I realized that the pin list could be rolled into this file because it's generic too