Added a slave select line to the SPI peripheral
This is to have the bare minimum needed to transfer one byte to one slave. This will be sufficient for the majority of users.
Starting adding an SPI manager which will give me the ability to manager multiple slaves and slave select lines.
In this case, the SPI peripheral and slave select are based on the SPISlave object. That way, calling functions do not need to care which peripheral goes to which device, and neither do any GPIO functions.