Beginning work on SPI manager
This file will maintain a list of SPI devices and their associated peripeherals. The SPI devices are referenced by name and each one has a pointer to it's own peripheral, so really there's no limit to how many devices you add.
It will go through the list of devices and check if a transfer needs to be performed. The user will have to handle the size of the data arrays. I did this so that there is no limit to how much data can be transferred. The SPI device will have a pointer to its buffers. Just call the transfer function and give it the amount of bytes you want to send and receive.
If you do have more than one slave device on a peripheral, you will have to have a function to handle the slave select pins. I am adding a callback function to help with this.