State machine is shaping up nicely
Right now, I'm only going to deal with master mode. This is simpler because after every transmit there is a receive. We don't need to worry about asynchronous bytes coming in. And it's the most common use case.
Added state called "request start". Modified function to check if a peripheral is busy. This way, I don't always have to traverse the list. I can first see if a transfer request has happened, then check if the peripheral is in use already.