Basic interface is ready. The SPI manager is still in the design phase.
There's still a lot to do with the SPI manager. I think I will make it a little different that the ADC manager. With the ADC and GPIO manager I placed the ADC and GPIO objects as external. But the SPI manager needs to manage both peripherals and slave devices. In this case, I think I will make it so that SPI devices are added to the manager by whatever .c file calls it and not have the peripherals as extern. On the other hand, with the other managers I expected them to be edited based on the project. There's no reason I couldn't do the same with the SPI one.