I changed the SPI Manager to be an object. Now each peripheral can have one manager.
Instead of trying to make one huge file do everything, I can divide and conquer. Each manager is created with an SPI peripheral, then SPI slaves are added to the manager. So now, each SPI manager only has to deal with one list. Each slave object had to be linked to peripheral anyway, so this just simplifies the whole process. SPI peripherals are still dereferenced using their interfaces and there are no hardware specific dependencies in the SPI manager file. So you can have as many or as few as you like.