Added transmit finished function. Renamed transmit finished event
I changed the old transmit finished event function to be transmit register empty event. This function is exactly the same as it used to be. I need to add a new function to check if the transmit is finished. This is slightly different as it checks the actual status of the UART shift register. Most of time, this function is not needed. It's possible that some microcontrollers may not even have this feature. In that case, it should be the same as the transmit register empty function. This extra function is for the rare case that we need to know exactly when the transmission is finished in order to set a GPIO or something. In my case, I only needed to use it for the very last byte of a data packet.