Doxygen. Working on compare match interrupts.
I want to use a channel number for compare match so that it is generic and can be used with different processors. The only issue is how the interrupts work are different for each processor. The best I can do is just number the compare match channels. The user will have to match up the number when calling the event handler. It's not great, but it does work. Although, if the processor is a multi vector interrupt style, it might be possible to simply place the interrupt in the library itself.