Changed how the initialization works. Made a STM32 generic subclass
Still trying to figure out exactly how I want to do this. I think the inclusion of the subclass is needed in case I want to make some non-interface functions in the future. Like for example, PWM related functions that set pins direclty on the STM32 or change the polarity.