Comments. Update compare match.
The compare match event function gets called for every compare match interrupt regardless of where it is. Then the function goes through and looks at the individual flags, clears them, and performs the callback.
Now the user doesn't have to worry about matching up the numbers on the compare match interrupt with the numbers in the timer. The library handles that.