New delay functions for STM32 fully tested.
The ms delay is very accurate. The us delay is close. It will probably vary a little depending on compiler optimizations etc. Delaying 1 or 2 us isn't as accurate as 10 to 100.
The 10 us and 100 us delay came to 10.57 us and 100.57 us after adjusting the value that is subtracted from period1us.