I switched up how the scheduler works
Before it would simply set a flag every tick. Then when the flag is set, it updates the counters and executes function calls. If a single task took longer than the shortest period, the task would not get executed until the next tick.
Now instead of single counter, each task has a counter that gets ticked every interrupt. When a task is ready it sets a pending flag.
Now I can look at tasks and see if multiple tasks are pending at once and decide what to do. This lets me add a priority number to the task as well.
Every task must still run to completion though, so care must be taken to set the periods and priorities.