Switched it up again.
Now instead of just a pending flag, I have a pending list and a addToPending flag. So in reality there are two lists. One that holds the list of all tasks, and one that holds the list of tasks to execute.
I made it so that when a pending task needs to get added to the list, it can't skip all the way to the front of the line. This will hopefully make it so that if there is a high priority task pending every loop it has to alternate with whoever is in the list. I'm hoping this will prevent starving low priority tasks. But I'm completely new at this, so I don't know how it will work yet. Everything I've found online is just ways to create a symbolic task in C++, and compute wait times and burst times for it using different algorithms. No one actually explains how the fucking thing is supposed to actually work.

So the way mine works is this: Every interrupt, go through the list and quickly tick every task timer. Set addToPending list flags.
Then in the main loop, run through the list and add pending tasks to a new list
When tasks are added to the pending lists they are inserted based on priority
If there is another task in the list with equal priority, the new one is inserted after
Go through the pending list and execute each function, then clear the pending flag
The addToPending flag should prevent a task from being added to the pending list if it already in the list.