Changed variadic function a little bit
The va_arg function promotes uint8_t to int, so I get a compiler warning. It will still work because it will convert to an uint8_t afterwards. But if the user passes in a -1, it will be converted to FF.
It's best to check the argument first. Also I don't like compiler warnings