Fixed output of comparator
It was getting stuck inbetween levels and not going to the upper level. I was not using the correct threshold for the upper and lower bounds. I kept trying different methods like looking at the previous input, previous output etc. until I realized the solution is much simpler. If you are currently in a middle output level, your thresholds are not the upper and lower thresholds. They are the upper + 1 upper threshold, and lower - 1 lower threshold. I set the default values to 0 and 0xFFFF that way it simplifies the logic a huge amount. If we are in the uppermost output level, the upper threshold check gets skipped because the input can never be greater than 0xFFFF. Similarly, for the lower threshold.