Cleaned up comparator
Now that I've fixed how the comparator library works, overlapping thresholds don't really matter anymore. So I removed the code that attempted to fix the user's array. It's not good to try and assume what the user wants. It's better to assume the user is right. If they want to have two deadzones with an overlap, they can now. The outer thresholds are used if it's in the middle.