Added simple min and max functions
A preprocessor macro will fail if someone were to do something like MIN(x++, y++). We all say we know not to do it, but the person who looks at your code years later may not know any better.
Added some default settings for the clock initialization for the F1