Thinking of ideas for new library
I don't like the fact that most emulated eeprom libraries I've looked at use an array to hold entries. The array is simple, but you have to create it with a known size. Meaning you have to determine how many entries you want to hold up front. I also don't like that you have to use a word to hold every single byte. It's inefficient. So I plan to make mine with different types of formats for more flexibility. The first four bytes of a page in flash should hold a header which has information about the status of the page and the format of the data being held.