Cleaned up STM32F1 UARTs. Finished G0 implementation
I changed the defines to not have the number prefixed. For some reason, I was thinking the names of the defines would conflict. But the defines are in the .c file, not the header file. This will make it much, much easier to make new implementations in the future now.