I think I finally got the convert to fixed point the way I want it. Still needs testing
It lets you input a 16-bit fractional part. Since the maximum is 65535, this would give you up to 5 digits of precision. You have to include the amount of precision for the fractional part as well as the number of fractional bits you want.
Entering a number like "3" with a precision of 3 is equivalent to saying "0.003". "55" with a precision of 3 would be "0.055".
Entering less precsion than the size of your fractional part will make it round. "65535" precision 4 = "6553". "65535" precision 3 = "655"