Added type specifier.
Becuase I know the product of two 16 bit numbers needs to be a 32 bit number, I added a width type, and changed the value to be 32 bit. This is going to take me some time to decide how to do these functions now. They were originally returning the Fxp type directly. I may change this to pass in a pointer to where the result should go... I'm not sure yet.