Changed behavior if displaying multiple error codes
I previously had it only insert a short pause if the current error code was 1. This makes the LED blink at a steady rate. I later added the ability to flash mulitple error codes. If someone had error code 1 and 2 active, there wasn't a very long pause between the two codes. I added a display active mask which masks off the number of codes that we want to display. Then I check to see if this mask is equal to error code 1.