Added some checks in FXP multiply and divide.
If the multiplication operands are bigger than 16 bit, shift them right first. If the dividend is already bigger than 16 bit, don't shift it right first. Instead figure how far we can shift left. Usually the numbers will be limited to 16 bits. Unless the previous result is from a multiplication. So I first check if the number is greater than 16-bits in order to skip this step and speed up the code.