Added my number to ascii function as just a utility function
It's a simple function that can save a lot of memory if you just need to convert a few numbers and don't need the full sprintf functionality.
I improved the left and right justified negative number section a little bit. If the number is right justified and there is space, it simply places a negative sign. Otherwise, it pushes the number string to the right.