SPI implementation should be ready to test
The good news is that it appears to pretty much work just like UART in terms of data registers, interrupts, etc. The only difference is the receive interrupt can't fire unless we transmit, but that makes no difference in terms of the transmit and receive code itself. The functions still operate the same way regardless.