Changed variable name. Added ability to send or receive zero bytes
There's no reason why the SPI can't have the readCount be zero. There are lot of situations where the SPI master is only using master out with no master in. I'm already checking the readWriteCount before placing data in the read buffer anyways. So if it's zero it just gets skipped over.