Added boolean to SPI manager. Removed init
The SPI manager can't call the init function. It has no idea what parameters are needed. And it shouldn't need to. The peripherals need to be configured beforehand. This was an idea left over from the ADC manager. There was only one ADC at the time, so I thought it would be more convenient to just have the manager handle the initialization. I realize now, that's not as good an idea as I thought it was. I should go back and change my ADC manager to be more like my new SPI manager instead. Where each manager is one object and each object is assigned to a peripheral.