The include is not needed
Unlike the G0, the declaration of the STM32 typedefs are actually in map.h. However, they are behind another define called "_USART1". This define goes to another file called stm32f10x_conf.h which is the project folder. No idea why they did this on the F1. It's prett dumb.
I may go back and change my "USART1" and "SPI1" defines to go to the address directly, so that it still compiles with just the map.h file. But I will want to add some notes to explain what's going on.