Removed RXNE check
I've had a lot of issues with the RXNE flag on this F1 processor. Something is clearing the flag automatically right after transmit. It does work, but only if you check the RXNE flag immediately after calling the transmit function.
I'm skeptical of it, so I've removed it for now, pending some more testing. In master mode, TXE should be sufficient.