Working on changes to GPIO library.
I'm trying to simplify the GPIO library. There is only one GPIO peripheral, so we don't really need to have instances of the GPIO object. Referring to the pin by name was cool and all, but this is much simpler.
I've combined the port and pin number into a single value. The upper nibble is the port and the lower nibble is the pin number in that port.
Now I can remove the pointer to the GPIO object and just replace it with this combined port pin value. The implementation can handle decoding it. (which is not difficult)
I can still refer to the pins with define statements like I normally used to do. Except now I don't need to define "port , pin" like I would have done had I used separate arguments.
Also, when the two arguments are separated, people tend to make copy and paste errors. Like copying and pasting lines of code and changing the pin, but forgetting to change the port.
I still needed a way to use inheritance, but I realized it's only necessary for the setup of the pin. So the create init type function is in its own section of the header file. This function still gets called with the sub class and still uses typecasting.
The other functions have to be implemented exactly as written for that processor. This is much simpler now that here is no sub class to set the actual pin anymore and the port and pin are combined.
I could have still written a .c file to call these functions like I do my other interface for the UART, but I wanted to keep the GPIO functions simple and fast by calling them directly.
So the GPIO library is sort of combination of my other two topologies, the interface with inheritance and the simple interface.