Finished prescale calculator. Added changes to F1 and G0
The period calculation function should be calculating based on the desired frequency.
The timer for the F1 uses a prescale counter. What I did was to convert the clock value to a 27.5 fixed point number in order to compute the desired prescale counter value.
I left my original method in the function too for reference. It goes through powers of two and finds the prescale value that gets us closest to the maximum timer value.