More changes to PWM
Enable compare is now EnableComparePWM. It seems that most microcontrollers don't have a separate enable and disable control. They usually just have a mode setting and sometimes an output pin setting. Others tend to just leave the output enable and disable to the port control registers. For now , I'm only going to be using PWM so I changed the name of the function. This is all most people care about anyway. In the future, EnablePWM will probably turn into SetMode or something.