Figured out how to do type punning. Tested
This technically works, although it's not super safe. It requires the first member of the sub class struct to be a pointer to the super class. I added example code and notes about how it works.
The safest method is still to make a sub class create function and use that. This is what I was doing before. I just forgot that it was necessary and not optional. I just using it for convenience.
Need to update all my notes to make sure to mention not to forget to set the super class pointer.
The example subclass Foo_MCU1 shows the proper way. It's just that the notes made it seem like it was an option. When it's not.