PWM is working!
Tested both set 16-bit and set percent functions.
The set compare 16-bit function don't have to be scaled. That was for the duty cycle function. It was overflowing the result variable. I shouldn't really need to scale the numbers anyway. The compute function is designed to make the period as large as possible. And two 16-bit numbers multiplied will yield a 32-bit number.