FIxed Park Miller. Changed left
The upper boundry for the LCG is 2^63. I think I need to use bits 62 through 31. I may have to change it to bits 62 through 32. We'll see how it looks.
The seed of the Park Miller can be changed because if m is a prime number, any number between 1 and m-1 will work.