Added comma separated output. Option to switch PM output.
Spectral test looks pretty good. The LCG matches my previous test when I compared the full output of an LCG vs just the upper bits, so I'm pretty confident it's implemented correctly. The Park Miller is the same as the one used in C++, and it also looks similar to the LCG. So that's good.