Added Schrage version.
I tested the output of the Schrage against the 32-bit Park Miller and it was identical.
I will probably combine all these types of PRNG's into a single class at some point. Because the bound function is almost identical with the exception of the maximum range. So is the seed function.