Working on creating classes
There's only one problem at the moment. The "big" LCG returns 32-bit, therefore the "small" one returns 16-bit. The big one uses a 64-bit state variable and the small one uses 32-bit. Makes sense. But the Park Miller is more effecient. Because it is full cycle, it uses 32-bit state and returns a 32-bit result. So does it get its own functions or not?