Added LCG small. Tested
I found a multiplier from a research paper by L'ecuyer that looks like it yields a little better output than the typical C rand function. A small random number generator is no ideal, but I added this option so it could be implemented on a small micro and still make use of the logarithmic skip function in the future.