Rotary encoder update
I added an invalid number to the state table. My thinking was that if there ever is an invalid state transition the output number needs to be reset back to 0, so that the next valid transition will begin again on either +1 or -1. This should help when the rotary encoder has one detent per full cycle. Which means it would have to go through four separate states or counts, either positive or negative, for a valid click. Need to test this theory.