Experimenting with modulo bias removal method
It does actually work. The reason I couldn't tell was because I was using a 16-bit value. Which is good. There was hardly any bias at all.
In my example, I used the output of the PRNG then converted it to 8-bit, then performed modulo division again to give an output within a specified range. The bias seems to be the worst when output is limited to the range 0 to 99. It actually seems to get better as the range approaches 255 or moves away from it. The range 200 has a very small area with a bias. Lower ranges produce a flatter bias as well. Because all outputs become biased evenly again I guess.