Working on I2C interface
Built header file for I2C1. Right now it is still a mix of SPI and UART library ideas. That is mostly because the SPI does not have independent transmit and receive functions. The I2C sort of does, but not really. I haven't decided how to finalize everything yet. For now, I will keep it the way that my original PIC32 code had it.