Working on I2C manager
I originally had the slave devices in the SPI and I2C manager have a pointer to their manager device. I think I couldn't decide how I wanted to organize the slave devices. On one hand, I can do a very large list and each peripheral points to who it needs to send data to. On the other hand, the slave device shouldn't really need to know that kind of information. In another commit from a year ago or so I wrote that I intended to have one manager per peripheral. I probably already tried using a single manager for multiple peripherals and determined that there was too much going on in that single function.

The only time I can think of right now where it might be helpful to have the slave use that pointer to its manager is when requesting a data transfer. The user should ideally just want to write data to a slave device. They shouldn't have to worry about who that device belongs to. That's the whole point of the library and interface. Right now, I have the slave device checking the state before allowing data to be written and setting the request flag. But honestly that isn't really necessary. That's what the request flag is for anyway. Just write the data to the slave, set the request flag, and the manager should handle it.

I would have to change my function naming convention slightly though, because I want to write data to the slave to initiate the request for data transfer. Should it be _Manager_TransferDataSlave(*slave, ... ) or just SPISlave_BeginTransfer(*self, ... )?