I2C Manager
I'm probably going to be splitting the I2C slave object into a basic base class and one for the I2C manager that has the extra pointers for managing its list of devices.

Ideally I think I want the user/programmer to call the I2C slave in order to initiate a read or write sequence. The method that the MCC uses is to call the I2C peripheral (which is similar to my manager, except it is an explicit object. One per peripheral) and add a sequence of writes or reads to the peripheral. But this requires the person using it to already know which devices belong to which manager. Imagine two separate boards with the same temperature sensor, but on two different I2C peripherals. The user shouldn't have to care about which I2C manager to call. That's the whole purpose of the interface.

I also kind of like the idea of managing the I2C slave's buffers outside the slave. Maybe just put a pointer to the read and write buffers in the slave? But I'd have to know the buffer size or use a numOfBytesToRead like I'm doing right now