Saving progress. Made read and write to data request buffer
For now the buffer lives in the I2C slave object. That is because it is preferable to have at least a buffer of size 2. That will allow a master write and then a read at minimum. And rather than just making a note, I made it explicit. For now. I could change my mind.
I may end up just making the I2C slave a basic class and putting everything in the sub class. That gives the programmer/user more freedom to do whatever the hell they want.