I2CManager. Figuring out different functions and types
The I2CSlave will be the base class. It will be as generic as possible. I moved all the other variables to the I2C_Manager's slave type which I'm calling I2CSlave_Node for now.
So far there are a few basic functions which will need to be in the interface. SendDataRequest, GetDataRequest, IsDataTransferFinished, and GetDataTransferStatus.
I might make a "begin" data transfer function. Or just leave it up the sub class to decide how they want it done.