Work on I2C PIC implementation
Changed name of receive enable and disable to ReceiveByte and ReceiveByteCancel. The transmit function does not have an enable and disable. This makes sense because it's not like a UART where transmit and receive are independent. You just call the transmit function and it does the transmit. I want the receive function to be the same way. Instead of enabling it, calling receive byte implies that you call it and then it finishes receiving one byte and you have to call it again.