Added functions for getting data from slaves and data transfer reports
I prefer writing the data to the slave device. This is so the user/programmer doesn't need to worry about which peripheral or manager object the data goes to. I have a small buffer in the slave device that holds a pointer to the data and what type of transfer to request. Because the communication is synchronous I need to wait until the I2C manager is done before notifying the user. I also want to just check on the slave device to see if the data transfer is finished. It makes the whole process much simpler. The user doesn't care about where the data goes. They just write to the slave and check if the slave is finished. But I also wanted to be able to make room in the data transfer buffer to allow more data in while the peripheral/manager is processing the current data transfer request. I've added a few variables in the slave and a few static functions to help manage that. The I2C manager will get the data transfer request out of the slave and load it into its temporary buffer. The I2CSlave_ReadFromDataTransferBuffer will advance the tail of the buffer. When the manager is done and the transfer is finished, it will wirte a report back to the slave device. The I2CManager_GetFinishedTransferReport will create the report.