I2C manager debugging. Updated repeated start
Repeated start has to be held in the I2C manager object, not the event. Because once the state machine goes back to idle, it loses that old event which.
Repeated start does appear to be working according to my scope. But the I2C slave I'm using doesn't seem to be responding to it. Maybe it's too fast?