I2C implementation. Fixed baud rate.
The reference manual says that if there is no calculation mentioned in the datasheet, the default equation is used. (the default uses Fcy instead of Fcy/2). So you'd think since the I2C section does not mention anything about the baud rate, you should use the default equation right? Well you'd be wrong. You are supposed to use the equation with the clock divided by 2 in it.