Tested two I2C temperature sensors simultaneously with I2C manager
In this test I only loaded one data transfer into each I2C slave at a time, so the different I2C messages get interwoven by the I2C manager, but it worked beautifully. I've also tested loading more than one transfer into the buffer of one of the sensors. In that case, the I2C manager will perform a repeated start instead of a start then stop.

Right now I have to check the state of the I2C manager though, in order to keep the two devices from colliding. This is because even though the slave device is ready to transfer, it doesn't mean that the I2C manager is ready. The I2C manager could be busy with a different slave. Ideally, I'd prefer not to call I2C functions directly from slave devices. In order for us to check if the bus is busy, the slave could use a pointer to its I2C manager. Another option would be a get state function added to the slave functions. The other option is to just set the timeout timer to a very large value that would ensure the manager has had enough to time to go through all devices. I will probably change the "is slave ready" function to be a get buffer count function and add an additional get state function to keep track of where the slave is in the data transfer process.