Updates to I2C manager
Adding slave state and function to check the slave state. I'd like to have a better way for the slave to know when its data transfer starts. This is because the original function which was called IsSlaveReadyForDataTransfer doesn't really give us enough information. It lets us know the slave is ready, but it's still possible that the I2C manager is busy servicing a different slave device. Sometimes, like in the case of my temperature sensor, we have to start a timer and then come back check on the I2C slave later. So for those applications it's nice to know when to start the timer instead of just using a very large timer and hoping for the best. At the moment, the I2C manager is writing the state into the I2C slave devices, but that doesn't mean it has to be implemented that way.

I also changed the IsReadyForDataTransfer function to just be Slave_GetDataTransferBufferCount. I also added read, and get size functions. These functions were already included, they were just static functions used by the I2C manager. I changed the I2CSlave_DataTransfer function to be WriteToDataTransferBuffer to compliment the exisiting read function. These changes make the slave buffer management more universal. And it also would allow me, in the future, to have the slave buffers be managed indirectly via pointer instead of having a fixed array inside each slave device.