I2C manager updates. Fixed bugs. Added new functions
Using the buffer count is not as useful when we want to put things in the transfer buffer. I don't necessarily want to know if the buffer count is zero so much as the buffer is not full. So I added code and functions from my other buffer library. I also added the IsNotEmpty function as well as a get buffer count function which could used in the future.