I2C manager. Changed event flags. Fixed more bugs.
Accidentally deleted the master's currentTransferFinished flag when moving around the slave's transferFinished flag. Fixed.

I decided to use an event flag for both start data transfer and finish data transfer rather than just relying on the single transferFinished flag. This ensures that both the start and fnish events cannot be missed accidentally if the user/programmer were only looking at the state of the I2C slave device. For example, I was looking for the I2C slave to not be idle in order to start my conversion timer for my temperature sensor. But that requires me to constantly check the flags of my device which limits how slow I can operate the tick function of my temperature sensor. Because the slave could go from idle to transferring data and back to idle between refreshes. Ideally this shouldn't happen because my tick rate of my temperature sensor is relatively fast, but I don't like imposing this limitation.