Adding get bus error function to I2C manager
So far it does detect when there is a bus collision. Unfortunately, sending a stop condition does not work because the slave has a hold of the SDA line still. At least that is what I think is going on. I'm going to try a few different things to see if I can figure out how to make it recover.