Made a generic transfer object and buffer
I originally made this for I2C slave devices, but then I realized it could also apply to SPI or any serial communication where a separate read and write are performed. For now, I will just make this a generic object. I might expand this be a generic slave (or target device) and change the file to TargetDevice.h.
I've added functions to handle a buffer of data transfer objects. This will be useful in the future so that the I2C slave and SPI slave don't have to have redundant code. It will also allow them to have non-static size which will be flexible. For example, right now my I2C slave object has a fixed data transfer buffer size of 3. But in the future I could make this any value I need. Or different devices could have different sized buffers based on their needs. For now, I will leave the data transfer array as a fixed size inside the I2CSlave.