Updated I2C manager to use data transfer object. Compiled.
I may end up making the data transfer object a class and make the I2C slave an implementation of it. I don't know yet. There are a lot of functions in I2CSlave that just directly call the data transfer buffer functions. On the other hand it's nice having the I2C slave separate, because there might be a case where we don't want to use the data transfer buffer.