Adding I2C target implementation. Changed names of files
I had to change the II2CTarget interface to something else. Because the functions in II2CTarget began with I2CTarget and so did my implemenation. I was considering changing it to be more generic anyway and it was becoming confusing. Haven't decided if I want the I2C implementation to be TargetDevice_I2C or just I2CTarget.