Working on I2C target implementation
Moved I2C variables from the old I2CTarget file to the new implementation. Still haven't really decided if I want the I2C and SPI to both be a subclass of one type, or just make them separate. The I2C manager will have to operate with the subclass type because it is required to get the 7 bit address etc. Whereas if the I2CTarget where the base class, then it is up to implementation to decide if I want to use a ring buffer, or a buffer of just size one. In that case, the I2C manager doesn't care. It would be calling using the base class target I2CTarget.