I2C Target. Saving progress.
Still working on how I want to implement everything. After looking at how the I2C manager code works, I've decided to keep the IsDataTransferStarted event. The manager will call the function whenever the data transfer starts. This is similar to how the TransmitRegisterEmpty function is called via the interrupt in the UART library. It actually kind of works the same. It's just that in the UART example, those event functions were called elsewhere, but in this I2C target code they have to be called by the I2C manager. They still have to be called either way.