I2CTarget init functions
Still haven't completely decided how I want the initialization functions organized. I like having a "set init type" function to initialize the members of the init type object be a non-interface function. This gives the user the freedom to set the members of the struct manually, or call the function. Whatever they prefer. I also kind of like the idea of including the 7-bit address in the arguments for either the SetInitTypeParams functions or even just the Init function. If this function is part of the function table, it ensures that the user includes the address somewhere in their own implementation.