New I2C manager and target code work!
Surprisingly my double pointer code worked correctly on the first try. The actual problem was that I forgot to update the I2C manager with the new the read from I2C target function. Originally it returned an error code, where 0 meant no error. Then later I changed it to just set the length to 0 instead, but forgot to change the following if-statement.