I2C updates. Added a clear transfer started flag function
I wasn't sure if I really needed this function or not, but I couldn't make up my mind. In a system like a UART this would similar to the "transmit register not empty" event or interrupt flag which get set anytime the transmit register is empty, which usually happens right after you write to it. On the UART the data transmission begins automatically though. Whereas on a system like this it can't be automatic. I thought about whether or not to clear the transfer started flag after reading data out of the transfer buffer but couldn't make up my mind. Again, just because data was taken out of the buffer, doesn't necessarily mean the transfer has begun. Just like I found before in my previous experiments, ultimately, the I2C manager state machine is handling the data so I had to wait for this transfer started event in order to start my timer for my temperature sensor.