Updated baud rate equation
The equation I had originally was technically correct. It's just that the input to the peripheral is divided by 2. It's not obvious from datasheet, but after looking at the reference manual for the UART I realized it is the same as the I2C clock. Both use Fp which is the instruction clock.